earning_report_analyst_agent/requirements.txt
langgraph
langchain
langchain_openai
langchain-community
langchain-google-genai
langchain_huggingface
ollama
pandas
datetime
ipython
python-dotenv
matplotlib
sec_edgar_downloader
beautifulsoup4
chromadb
fuzzywuzzy
protobuf==3.20.3
